remove old nextjs example
