--- examples/js/nextjs-new/.dockerignore
@@ -1,2 +1,0 @@
-node_modules
-.next
