--- examples/js/nextjs-new/pages/_app.js
@@ -1,8 +1,0 @@
-import { start } from '../lib/bugsnag'
-
-start();
-
-export default function App({ Component, pageProps, err }) {
-  // Workaround for https://github.com/vercel/next.js/issues/8592
-  return <Component {...pageProps} err={err} />
-}
