+++ examples/js/nextjs/package.json
@@ -1,17 +1,20 @@
 {
-  "name": "bugsnag-js-example",
-  "version": "1.0.0",
-  "main": "app.js",
+  "name": "bugsnag-example",
+  "version": "0.0.1",
+  "private": true,
   "scripts": {
     "dev": "next",
     "build": "next build",
     "start": "next start"
   },
   "dependencies": {
-    "@bugsnag/js": "^7.0.0",
-    "@bugsnag/plugin-react": "^7.0.0",
-    "next": "^7.0.1",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "@bugsnag/js": "^7.9.2",
+    "@bugsnag/plugin-aws-lambda": "^7.9.2",
+    "next": "^10.0.8 || 11.x",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  },
+  "devDependencies": {
+    "webpack-bugsnag-plugins": "^1.6.0"
   }
 }
