+++ examples/js/nextjs/pages/_app.js
@@ -1,30 +1,8 @@
-import React from 'react'
-import App, { Container } from 'next/app'
-import Bugsnag from '../lib/bugsnag'
-import Error from './_error'
+import { start } from '../lib/bugsnag'
 
-const ErrorBoundary = Bugsnag.getPlugin('react')
+start();
 
-export default class MyApp extends App {
-  static async getInitialProps ({ Component, router, ctx }) {
-    let pageProps = {}
-
-    if (Component.getInitialProps) {
-      pageProps = await Component.getInitialProps(ctx)
-    }
-
-    return { pageProps }
-  }
-
-  render () {
-    const { Component, pageProps } = this.props
-
-    return (
-      <ErrorBoundary FallbackComponent={Error}>
-        <Container>
-          <Component {...pageProps} />
-        </Container>
-      </ErrorBoundary>
-    )
-  }
+export default function App({ Component, pageProps, err }) {
+  // Workaround for https://github.com/vercel/next.js/issues/8592
+  return <Component {...pageProps} err={err} />
 }
