+++ examples/js/nextjs/pages/api/scenario1.js
