+++ examples/js/nextjs/pages/api/scenario2.js
