+++ examples/js/nextjs/pages/api/scenario3.js
