+++ examples/js/nextjs/pages/api/scenario4.js
