+++ examples/js/nextjs/pages/client/scenario1.js
