+++ examples/js/nextjs/pages/client/scenario2.js
