+++ examples/js/nextjs/pages/client/scenario3.js
