+++ examples/js/nextjs/pages/client/scenario4.js
