+++ examples/js/nextjs/pages/client/scenario5.js
