+++ examples/js/nextjs/pages/ssr/scenario1.js
