+++ examples/js/nextjs/pages/ssr/scenario2.js
