+++ examples/js/nextjs/pages/ssr/scenario3.js
