+++ examples/js/nextjs/pages/ssr/scenario4.js
