+++ examples/js/nextjs/public/favicon.ico
